pyproject.toml
[tool.poetry]
name = "Reporting-Tool-Frontend"
version = "0.1.0"
description = "Your project description"
authors = ["Saurabh Lal <saurabh.lal@trusttyreporter.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
Flask-Mail = "^0.10.0"
authlib = "^1.0.0"
requests = "^2.31.0"
bcrypt = "^4.1.3"
psycopg2-binary = "^2.9.9"
Jinja2 = "^3.1.4"
Werkzeug = "^3.0.3"
pandas = "^2.2.2"
chardet = "^5.2.0"
itsdangerous = "^2.2.0"
openpyxl = "3.1.3"
langchain = "0.2.11"  # Specific version provided
langserve = "0.2.2"  # Specific version provided
fastapi = "^0.111.0"  # Specific version provided
"Flask-Ext" = "^0.1"  # Specific version provided
flask-session = "^0.8.0"  # Adjust version if necessary
redis = "^5.0.4"  # Specific version provided
fastapi-cli = "^0.0.3"  # Missing from initial list, added here
langchain-community = "0.2.10"  # Missing from initial list, added here
langchain-core = "0.2.23"  # Missing from initial list, added here
langchain-text-splitters = "0.2.2"  # Missing from initial list, added here
api-analytics = {extras = ["flask"], version = "*"}
httpx = "*"
aiofiles = "*"
flask-moment = "*"
celery = "*"
flask-sse = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
